refactor(frontend): 页面级组件移至 views/，组件与视图职责分离
将 auth（AuthLayout/LoginView/RegisterView）、Dashboard、ErrorBank、
ReviewView、ChatView、ChatPage、NoteView、SettingsView、SplitHistory
从 components/ 移至 views/，清理空目录。
