+++ frontend/src/views/DashboardView.vue
@@ -3,7 +3,7 @@ import { ref, watch, nextTick, onBeforeUnmount, computed } from 'vue'
 import * as api from '@/api.js'
 import CustomSelect from '@/components/ui/CustomSelect.vue'
 import GlassCard from '@/components/ui/GlassCard.vue'
-import StatCard from './StatCard.vue'
+import StatCard from '@/components/dashboard/StatCard.vue'
 
 const props = defineProps({
   theme: { type: String, default: 'light' },
