+++ frontend/src/views/ErrorBankView.vue
