+++ frontend/src/views/NoteView.vue
