+++ frontend/src/views/ReviewView.vue
@@ -3,7 +3,7 @@ import { ref, reactive, watch, nextTick, onBeforeUnmount, computed } from 'vue'
 import * as api from '@/api.js'
 import { typesetMath as _typesetMath } from '@/utils.js'
 import QuestionDetailModal from '@/components/question/QuestionDetailModal.vue'
-import AiAnalysisModal from './AiAnalysisModal.vue'
+import AiAnalysisModal from '@/components/review/AiAnalysisModal.vue'
 import CustomSelect from '@/components/ui/CustomSelect.vue'
 import PageHeader from '@/components/ui/PageHeader.vue'
 import QuestionItem from '@/components/question/QuestionItem.vue'
