+++ frontend/src/views/SettingsView.vue
@@ -2,8 +2,8 @@
 import { ref, inject, onMounted, watch } from 'vue'
 import { fetchAppConfig, updateAppConfig } from '@/api.js'
 import { genId } from '@/utils.js'
-import ProviderDialog from './ProviderDialog.vue'
-import ProviderSection from './ProviderSection.vue'
+import ProviderDialog from '@/components/settings/ProviderDialog.vue'
+import ProviderSection from '@/components/settings/ProviderSection.vue'
 
 const props = defineProps({
   visible: { type: Boolean, default: false },
