+++ frontend/src/views/SplitHistoryView.vue
