+++ frontend/src/views/WorkspaceView.vue
@@ -20,14 +20,14 @@ import ErasePreview from '@/components/workspace/ErasePreview.vue'
 import OcrPreview from '@/components/workspace/OcrPreview.vue'
 import ImageModal from '@/components/ui/ImageModal.vue'
 import ToastContainer from '@/components/ui/ToastContainer.vue'
-import Dashboard from '@/components/dashboard/Dashboard.vue'
-import ReviewView from '@/components/review/ReviewView.vue'
-import ErrorBank from '@/components/errorbank/ErrorBank.vue'
-import ChatView from '@/components/chat/ChatView.vue'
-import SettingsView from '@/components/settings/SettingsView.vue'
-import SplitHistory from '@/components/workspace/SplitHistory.vue'
-import NoteView from '@/components/note/NoteView.vue'
-import ChatPage from '@/components/chat/ChatPage.vue'
+import Dashboard from '@/views/DashboardView.vue'
+import ReviewView from '@/views/ReviewView.vue'
+import ErrorBank from '@/views/ErrorBankView.vue'
+import ChatView from '@/views/ChatView.vue'
+import SettingsView from '@/views/SettingsView.vue'
+import SplitHistory from '@/views/SplitHistoryView.vue'
+import NoteView from '@/views/NoteView.vue'
+import ChatPage from '@/views/ChatPageView.vue'
 
 // ---- 认证状态 ----
 const { currentUser } = useAuth()
