+++ frontend/src/views/auth/RegisterView.vue
